pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spirrow-magickit"
version = "0.1.0"
description = "Orchestration layer for Spirrow Platform"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Spirrow Team" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0.1",
    # Phase 2: Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Phase 2: WebSocket
    "websockets>=12.0",
    # Phase 2: WebUI
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
magickit = "magickit.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/magickit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
